Fishing Rod
Added one under player, but not in the camera. Idk, it just feels better when it moves with the player, but not the camera itself.
Dunno if ya'll read these, but if lets check: Could one of you code masterminds cap how far the character can look down? Just have them not be able to look directly on the ground, so they don't see their feet and the fishing rod sliding in the air when they have it enabled